+++ sdk/src/wallet/account/operations/syncing/addresses/output_ids/basic.rs
@@ -11,7 +11,7 @@ use futures::FutureExt;
 use crate::types::api::plugins::indexer::OutputIdsResponse;
 use crate::{
     client::{node_api::indexer::query_parameters::QueryParameter, secret::SecretManage},
-    types::block::{address::Bech32AddressLike, output::OutputId},
+    types::block::{address::Bech32Address, output::OutputId, ConvertTo},
     wallet::Account,
 };
 
@@ -22,9 +22,9 @@ where
     /// Returns output ids of basic outputs that have only the address unlock condition
     pub(crate) async fn get_basic_output_ids_with_address_unlock_condition_only(
         &self,
-        bech32_address: impl Bech32AddressLike,
+        bech32_address: impl ConvertTo<Bech32Address>,
     ) -> crate::client::Result<Vec<OutputId>> {
-        let bech32_address = bech32_address.to_bech32()?;
+        let bech32_address = bech32_address.convert()?;
         // Only request basic outputs with `AddressUnlockCondition` only
         Ok(self
             .client()
@@ -42,9 +42,9 @@ where
     /// `ExpirationUnlockCondition` or `StorageDepositReturnUnlockCondition`
     pub(crate) async fn get_basic_output_ids_with_any_unlock_condition(
         &self,
-        bech32_address: impl Bech32AddressLike,
+        bech32_address: impl ConvertTo<Bech32Address>,
     ) -> crate::wallet::Result<Vec<OutputId>> {
-        let bech32_address = bech32_address.to_bech32()?;
+        let bech32_address = bech32_address.convert()?;
         // aliases and foundries
         #[cfg(target_family = "wasm")]
         {